moved emp-manager from /cpp to /c
