--- c/employee-manager/.gitignore
@@ -1,1 +1,0 @@
-*.swp
