+++ c/employee-manager/notes/architecture.md
