+++ c/employee-manager/src/headers/employee.h
@@ -1,20 +1,28 @@
-typedef struct {
-	char* uuid;
-  int age;
-  char* first;
-  char* last;
-  char* phone;
-  char* email;
-	char gender;
-} Employee;
+#ifndef EMPLOYEE_H
+#define EMPLOYEE_H
 
-Employee create_employee(
-	int age,
-	char* first,
-	char* last,
-	char* phone,
-	char* email,
-	char gender
-);
+#include <string.h>
 
-char* create_uuid();
+typedef struct e {
+	const unsigned long int employee_id;
+	std::string first_name;
+	std::string middle_name;
+	std::string last_name;
+	std::string suffix;
+	std::string email_address;
+	std::string address;
+	std::string home_phone_number;
+	std::string cell_phone_number;
+	std::string marital_status;
+	std::string social_security_number;
+	unsigned long long int start_date;
+	std::string emergency_contact;
+	std::string gender;
+	std::string ethnicity;
+	std::string residency;
+	std::string job_title;
+	unsigned long int salary;
+} employee;
+
+#endif
+
